Update code to fit coding standard
