+++ src/main/java/duke/command/Command.java
@@ -1,10 +1,10 @@
-package Duke.command;
+package duke.command;
 
-import Duke.exception.DukeException;
+import duke.exception.DukeException;
 
-import Duke.util.TaskList;
-import Duke.util.Storage;
-import Duke.util.Ui;
+import duke.util.TaskList;
+import duke.util.Storage;
+import duke.util.Ui;
 
 /**
  * This Command class is the parent class for all commands to be executed.
