+++ src/main/java/duke/util/Ui.java
@@ -1,4 +1,4 @@
-package Duke.util;
+package duke.util;
 
 import java.io.PrintStream;
