+++ src/test/java/duke/task/TaskStub.java
@@ -1,4 +1,4 @@
-package Duke.task;
+package duke.task;
 
 public class TaskStub extends Task {
